Refactor out ModelStub class from AddCommandTest class
